+++ hpx/concurrent/unordered_map.hpp
@@ -9,8 +9,7 @@
 #define HPX_CONCURRENT_UNORDERED_MAP
 
 #include <hpx/lcos/local/mutex.hpp>
-#include <plugins/parcelport/readers_writers_mutex.hpp>
-#include <plugins/parcelport/libfabric/rdma_locks.hpp>
+#include <hpx/lcos/local/readers_writers_mutex.hpp>
 //
 #include <boost/thread/locks.hpp>
 //
