+++ plugins/parcelport/verbs/rdma/rdma_controller.cpp
@@ -10,7 +10,7 @@
 #include <hpx/config/parcelport_verbs_defines.hpp>
 #include <hpx/config/parcelport_defines.hpp>
 //
-#include <plugins/parcelport/readers_writers_mutex.hpp>
+#include <hpx/lcos/local/readers_writers_mutex.hpp>
 //
 #include <plugins/parcelport/verbs/rdma/rdma_error.hpp>
 #include <plugins/parcelport/verbs/rdma/rdma_locks.hpp>
